refactor: Relocate create post related files to home package and use CreatePostViewModel only in CreatePost (#42)
- In `CategoryTagDrawer.kt`, pass `categoryList` as a parameter instead of observing it from the ViewModel.
- In `CreatePostViewModel.kt`, rename `_categoryTagList` and `categoryTagList` to `_selectedCategoryList` and `selectedCategoryList` respectively.
- In `CreatePostViewModel.kt`, change `getCategoryList()` to be called directly in `init` instead of launching a new coroutine.
