+++ app/src/main/java/com/example/rentit/feature/home/createpost/categorytag/CategoryTagDrawer.kt
@@ -1,4 +1,4 @@
-package com.example.rentit.feature.createpost
+package com.example.rentit.feature.home.createpost.categorytag
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -7,26 +7,22 @@ import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.example.rentit.common.component.screenHorizontalPadding
 import com.example.rentit.common.theme.RentItTheme
 import com.example.rentit.data.product.dto.CategoryDto
-import com.example.rentit.feature.createpost.component.TagButton
+import com.example.rentit.feature.home.createpost.LabeledContent
+import com.example.rentit.feature.home.createpost.component.TagButton
 
 @OptIn(ExperimentalLayoutApi::class)
 @Composable
 fun CategoryTagDrawer(
-    selectedCategoryList: List<CategoryDto>,
+    categoryList: List<CategoryDto>,
+    selectedCategoryList: List<CategoryDto> = emptyList(),
     onTagButtonClick: (CategoryDto) -> Unit
 ) {
-    val createPostViewModel: CreatePostViewModel = hiltViewModel()
-    val categoryList by createPostViewModel.categoryList.collectAsStateWithLifecycle()
-
     val parentCatList = categoryList.filter { it.isParent }
     Column(
         modifier = Modifier
@@ -58,6 +54,6 @@ fun CategoryTagDrawer(
 @Composable
 fun PreviewCategoryTagDrawer() {
     RentItTheme {
-        CategoryTagDrawer(selectedCategoryList = emptyList()){}
+        CategoryTagDrawer(categoryList = emptyList()){}
     }
 }
